floating point imprecision

+++ CS50C-2024/calc.c
@@ -8,7 +8,7 @@ int main(void)
     int x = get_int("What's x ?; ");
     int y = get_int("What's y ?; ");
 
-    printf("%f\n", (float) x / (float)y);
+    printf("%.20f\n", (double) x / (double)y);
 
 }
